Finished working on 3rd test case for 3rd User Story
