+++ src/test/java/com/trycloud/fatih/practice/US2.java
@@ -13,6 +13,17 @@ import java.util.List;
 public class US2 extends TestBase {
 
 
+    @Test
+    public void loginValid(){
+        driver.get(url);
+        driver.findElement(By.id("user")).sendKeys(username1);
+        driver.findElement(By.id("password")).sendKeys(password);
+        driver.findElement(By.id("submit-form")).click();
+        Assert.assertEquals(driver.getCurrentUrl(),"http://qa.trycloud.net/index.php/apps/dashboard/");
+    }
+
+
+
 
     @Test
     public void mainModules(){
