+++ src/test/java/com/trycloud/tests/base/TestBase.java
@@ -2,14 +2,11 @@ package com.trycloud.tests.base;
 
 import com.trycloud.utilities.ConfigurationReader;
 import com.trycloud.utilities.WebDriverFactory;
-import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
-import org.testng.Assert;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
-import org.testng.annotations.Test;
 
-public abstract class TestBase {
+    public abstract class TestBase {
         //This class created for re-usability of methods
 
         //               Data for TryCloud WebSite :
@@ -30,15 +27,6 @@ public abstract class TestBase {
             driver = WebDriverFactory.getDriver(ConfigurationReader.getProperty("browser"));
             assert driver != null; // it is just to get rid of warning
             WebDriverFactory.manageWebDriver(driver);
-        }
-
-        @Test
-        public void loginValid(){
-            driver.get(url);
-            driver.findElement(By.id("user")).sendKeys(username1);
-            driver.findElement(By.id("password")).sendKeys(password);
-            driver.findElement(By.id("submit-form")).click();
-            Assert.assertEquals(driver.getCurrentUrl(),"http://qa.trycloud.net/index.php/apps/dashboard/");
         }
 
         @AfterClass // this will run 1 time only after class
